+++ consensus-proof/src/checkpoint.rs
@@ -1,3 +1,1 @@
-pub fn test_checkpoint() {
-    println!("testing from checkpoint...");
-}
+
